requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0
django-tinymce==3.4.0
mysql-connector-python==8.0.27
mysqlclient==2.1.0
protobuf==3.19.3
psycopg2==2.9.3
PyMySQL==1.0.2
PyYAML==6.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
